Update load_table_checks
Updated the load_table_checks function to pull from the defaults file for column labels instead of hardcoding as much, improved the error messages, and fixed some of the logic that was causing spurious warnings about columns that were not invalid.

May consider adding checks here that print warnings for duplicate columns in the same column group. For instance, if "Date" is in a protocol column group twice this should at least be a warning if not an error.